webapi/webapi-rawsockets-w3c-tests/rawsockets/TCPServerSocket_interface_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>

-->

<meta charset="utf-8">
<title>Raw Socket Test: TCPServerSocket interface basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/raw-sockets/#interface-TCPServerSocket">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>

var tcpserversocket = new namespace.TCPServerSocket(socketOptions);

[
  ["DOMString", "localAddress", "readonly"],
  ["unsigned short", "localPort", "readonly"],
  ["ReadyState", "readyState", "readonly"],
  ["event", "onconnect", ""],
  ["event", "onerror", ""],
  ["function", "close", ""],
  ["function", "suspend", ""],
  ["function", "resume", ""]
].forEach(function(attr) {
  var type = attr[0];
  var name = attr[1];
  var read = attr[2];

  test(function() {
    assert_true(name in tcpserversocket, name + " attribute in TCPServerSocket element");
    switch(type) {
      case "DOMString":
        assert_equals(typeof tcpserversocket[name], "string", name + " attribute of type");
        break;

      case "function":
        assert_equals(typeof tcpserversocket[name], "function", name + " attribute of type");
        break;

      case "ReadyState":
        assert_equals(typeof tcpserversocket[name], "string", name + " attribute of type");
        break;

      case "unsigned short":
        assert_equals(typeof tcpserversocket[name], "number", name + " attribute of type");
        var value = tcpserversocket[name];
        assert_true(0 <= value && value <= 65535, "unsigned short " + value + " not in range [0, 65535]");
        break;

      default:
        break;
    }

    switch(read) {
      case "readonly":
        assert_readonly(tcpserversocket, name, name + " attribute");
        break;

      default:
        break;
    }
  }, "Check if " + read + " TCPServerSocket." + name + " exists and type of " + type);
});

</script>
